Make BuiltinByteString (and others) more opaque
In the presence of bang patterns, GHC will very aggressively take apart
`BuiltinByteString`s and look into the underlying `ByteString`. This is
bad for our users.

We can avoid this by making `BuiltinByteString` `data` instead of
`newtype` (at the cost of writing some tedious instnaces). (Also
applies to the other `Builtin*` types that were supposed to be opaque.)

However, that doesn't save us entirely. GHC will still generate pattern
matches on `BuiltinByteString` to evaluate it. For some reason, it
doesn't generate default-only matches (which we already handle), even
though it doesn't use any of the variables that it binds.

Fortunatley, we can easily adapt the code to handle the more general
case of single-alternative case expressions which don't use any of the
bound variables.

(I scratched my head about why this wasn't working for a while until I
came to the realisation recorded in `Note [Occurrence analysis]`.)

As a bonus, this also generates slightly better code for the common case
of matching on `Unit`, which happens to generate such pattern matches
also!